almost there, there is only the viewer that seems broken but it is probably due to the name system that we need to talk about and rethink.

+++ autopack/Environment.py
@@ -1748,6 +1748,10 @@ h1 = Environment()
         """
         Save the current grid and the compartment grid information in a file. (pickle) 
         """
+        d = os.path.dirname(gridFileOut)
+        if not os.path.exists(d):
+            print ("gridfilename path problem",gridFileOut)
+            return
         f = open(gridFileOut, 'wb')#'w'
         self.writeArraysToFile(f) #save self.gridPtId and self.distToClosestSurf
